+++ node-commons/src/main/java/org/apache/iotdb/commons/sync/PipeStatus.java
@@ -16,8 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.commons.sync.pipe;
+package org.apache.iotdb.commons.sync;
 
+@Deprecated
 public enum PipeStatus {
 
   // a new pipe should be stop status
