+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/lagacy/IoTDBSyncReceiverV1_1.java
@@ -21,11 +21,7 @@ package org.apache.iotdb.db.pipe.connector.lagacy;
 
 import org.apache.iotdb.common.rpc.thrift.TSStatus;
 import org.apache.iotdb.commons.exception.IllegalPathException;
-import org.apache.iotdb.commons.exception.sync.PipeDataLoadException;
 import org.apache.iotdb.commons.path.PartialPath;
-import org.apache.iotdb.commons.sync.transport.SyncIdentityInfo;
-import org.apache.iotdb.commons.sync.utils.SyncConstant;
-import org.apache.iotdb.commons.sync.utils.SyncPathUtil;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.mpp.plan.Coordinator;
@@ -33,8 +29,12 @@ import org.apache.iotdb.db.mpp.plan.analyze.IPartitionFetcher;
 import org.apache.iotdb.db.mpp.plan.analyze.schema.ISchemaFetcher;
 import org.apache.iotdb.db.mpp.plan.execution.ExecutionResult;
 import org.apache.iotdb.db.mpp.plan.statement.metadata.DatabaseSchemaStatement;
+import org.apache.iotdb.db.pipe.connector.lagacy.exception.SyncDataLoadException;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.PipeData;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.TsFilePipeData;
+import org.apache.iotdb.db.pipe.connector.lagacy.transport.SyncIdentityInfo;
+import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncConstant;
+import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncPathUtil;
 import org.apache.iotdb.db.query.control.SessionManager;
 import org.apache.iotdb.rpc.RpcUtils;
 import org.apache.iotdb.rpc.TSStatusCode;
@@ -280,7 +280,7 @@ public class IoTDBSyncReceiverV1_1 {
       pipeData.createLoader().load();
       logger.info(
           "Load pipeData with serialize number {} successfully.", pipeData.getSerialNumber());
-    } catch (PipeDataLoadException e) {
+    } catch (SyncDataLoadException e) {
       logger.error("Fail to load pipeData because {}.", e.getMessage());
       return RpcUtils.getStatus(
           TSStatusCode.PIPESERVER_ERROR, "Fail to load pipeData because " + e.getMessage());
