+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/lagacy/pipedata/TsFilePipeData.java
@@ -20,11 +20,11 @@
 package org.apache.iotdb.db.pipe.connector.lagacy.pipedata;
 
 import org.apache.iotdb.commons.exception.IllegalPathException;
-import org.apache.iotdb.commons.sync.utils.SyncConstant;
 import org.apache.iotdb.db.engine.modification.ModificationFile;
 import org.apache.iotdb.db.engine.storagegroup.TsFileResource;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.load.ILoader;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.load.TsFileLoader;
+import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncConstant;
 import org.apache.iotdb.tsfile.utils.ReadWriteIOUtils;
 
 import org.slf4j.Logger;
