+++ server/src/main/java/org/apache/iotdb/db/pipe/connector/lagacy/pipedata/queue/BufferedPipeDataQueue.java
@@ -19,11 +19,11 @@
 package org.apache.iotdb.db.pipe.connector.lagacy.pipedata.queue;
 
 import org.apache.iotdb.commons.exception.IllegalPathException;
-import org.apache.iotdb.commons.sync.utils.SyncConstant;
-import org.apache.iotdb.commons.sync.utils.SyncPathUtil;
 import org.apache.iotdb.commons.utils.FileUtils;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.PipeData;
 import org.apache.iotdb.db.pipe.connector.lagacy.pipedata.TsFilePipeData;
+import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncConstant;
+import org.apache.iotdb.db.pipe.connector.lagacy.utils.SyncPathUtil;
 import org.apache.iotdb.tsfile.exception.NotImplementedException;
 
 import org.slf4j.Logger;
